requirements.txt
fastapi
uvicorn
python-multipart
requests
redis
python-json-logger
python-dotenv
